Refactor transaction controller to DRY dec code
