+++ test/config.js
@@ -13,19 +13,7 @@ class psuedoRes {
   }
 }
 
-function dec(stringifiedObject) {
-  if (typeof atob === 'undefined') {
-    let buffer = Buffer.from(stringifiedObject, 'base64');
-    return new Uint8Array(Array.prototype.slice.call(buffer, 0));
-  } else {
-    const decodedB64 = atob(stringifiedObject);
-    const arr = new Uint8Array(decodedB64.length);
-    for (let i = 0; i < decodedB64.length; i++) {
-      arr[i] = decodedB64.charCodeAt(i);
-    }
-    return arr;
-  }
-}
+const dec = require('../server/utils/dec');
 
 global.signObject = (messageObject, secretKey) => {
   const stringifiedObject = JSON.stringify(messageObject);
